Add ability to embed vtkjs in ParaView Glance file
Now, when "Export to VTK.js" is clicked in the OpenVR
panel, it gives you the option to embed the VTK.js file
in a ParaView Glance html file. This will make a copy of
the ParaView Glance html file and give it the same name
as the vtk.js file (with a `.html` extension), and embed
the vtk.js file inside it.

If cancel is pressed, the embedding will not be performed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>