Add export to VTK.js dialog for OpenVR
The dialog allows the user to choose some things they could already
choose before (the VTK.js file name and the optional ParaView glance
file name), and it also adds options for writing textures and writing
texture LODs.

Once accepted, the settings that the user chose are saved in the
pqSettings, and they will be restored when the dialog is executed again.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>