Add polyLOD options to pqExportToVtkJsDialog
This adds options for writing out polyLODs for use in VTK.js. In
addition, the base class variable "FilesToZip" are being zipped in
the vtkPVWebExporter as well, because it is already zipping the
.vtkjs file, and they were simple to add. Those files are the poly LOD
directories. Because of the new variable, this commit depends on the
corresponding commit in VTK that adds the "FilesToZip" variable.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>